refactor(repo): replacing StatsItem, KeyStatsItem component to ui-components
