--- apps/tangle-dapp/src/components/KeyStatsItem/KeyStatsItemVariant.ts
@@ -1,7 +1,0 @@
-enum KeyStatsItemVariant {
-  Left,
-  Center,
-  Right,
-}
-
-export default KeyStatsItemVariant;
