--- apps/tangle-dapp/src/components/KeyStatsItem/index.ts
@@ -1,2 +1,0 @@
-export * from './KeyStatsItem';
-export { default as KeyStatsItem } from './KeyStatsItem';
