+++ apps/tangle-dapp/src/components/index.ts
@@ -1,7 +1,6 @@
 export * from './BondedTokensBalanceInfo';
 export { default as CardWithTangleLogo } from './CardWithTangleLogo';
 export * from './HeaderChip';
-export * from './KeyStatsItem';
 export * from './NominatorStatsItem';
 export * from './Sidebar';
 export * from './skeleton';
