+++ apps/tangle-dapp/src/containers/KeyStatsContainer/ValidatorCountKeyStat.tsx
@@ -1,7 +1,7 @@
 import { EMPTY_VALUE_PLACEHOLDER } from '@tangle-network/ui-components/constants';
 import { FC } from 'react';
 
-import KeyStatsItem from '../../components/KeyStatsItem/KeyStatsItem';
+import { KeyStatsItem } from '@tangle-network/ui-components';
 import useValidatorCountSubscription from '../../data/KeyStats/useValidatorsCountSubscription';
 
 const ValidatorCountKeyStat: FC = () => {
