+++ libs/ui-components/src/components/Stats/index.ts
@@ -1,1 +1,4 @@
 export * from './Stats';
+export { StatsItem } from './StatsItem';
+export { KeyStatsItem } from './KeyStatsItem';
+export { KeyStatsItemVariant } from './types';
