+++ libs/ui-components/src/components/Stats/types.ts
@@ -1,3 +1,4 @@
+import { ReactNode } from 'react';
 import { PropsOf, IComponentBase } from '../../types';
 
 import { TitleWithInfoProps } from '../TitleWithInfo/types';
@@ -9,4 +10,38 @@ export interface StatItem {
 
 export interface StatsProps extends PropsOf<'div'>, IComponentBase {
   items: Array<StatItem>;
+}
+
+/**
+ * StatsItem types
+ */
+export type StatsItemProps = {
+  title: string | ReactNode;
+  tooltip?: string | ReactNode;
+  className?: string;
+  children: ReactNode | null;
+  isError: boolean;
+};
+
+/**
+ * KeyStatsItem types
+ */
+export type KeyStatsItemProps = {
+  title: string;
+  prefix?: string;
+  suffix?: string;
+  tooltip?: string;
+  className?: string;
+  showDataBeforeLoading?: boolean;
+  hideErrorNotification?: boolean;
+  children?: ReactNode;
+  error: Error | null;
+  isLoading?: boolean;
+  variant?: KeyStatsItemVariant;
+};
+
+export enum KeyStatsItemVariant {
+  Left,
+  Center,
+  Right,
 }
